renamed a main component file
