+++ project/src/components/main-page/main-page.tsx
